requirements.txt
pydantic==2.10.4
azure-storage-blob==12.24.0
adafruit-circuitpython-dht==4.0.6
RPi.GPIO==0.7.1
python-dotenv==1.0.1
